xero_backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
requests==2.32.3
pytest==8.3.2
PyYAML==6.0.2
